ZK ACL Patch for 8.6.3

+++ solr/core/src/test/org/apache/solr/cloud/api/collections/CollectionsAPIAsyncDistributedZkTest.java
@@ -40,6 +40,7 @@ import org.apache.solr.common.SolrException;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.cloud.Replica;
 import org.apache.solr.common.cloud.Slice;
+import org.apache.solr.common.cloud.ZkStateReader;
 import org.apache.solr.common.util.ExecutorUtil;
 import org.apache.solr.common.util.SolrNamedThreadFactory;
 import org.junit.After;
@@ -98,6 +99,13 @@ public class CollectionsAPIAsyncDistributedZkTest extends SolrCloudTestCase {
 
   @Test
   public void testAsyncRequests() throws Exception {
+    boolean legacy = random().nextBoolean();
+    if (legacy) {
+      CollectionAdminRequest.setClusterProperty(ZkStateReader.LEGACY_CLOUD, "true").process(cluster.getSolrClient());
+    } else {
+      CollectionAdminRequest.setClusterProperty(ZkStateReader.LEGACY_CLOUD, "false").process(cluster.getSolrClient());
+    }
+    
     final String collection = "testAsyncOperations";
     final CloudSolrClient client = cluster.getSolrClient();
 
@@ -206,9 +214,11 @@ public class CollectionsAPIAsyncDistributedZkTest extends SolrCloudTestCase {
       .processAndWait(client, MAX_TIMEOUT_SECONDS);
     assertSame("DeleteReplica did not complete", RequestStatusState.COMPLETED, state);
 
-    state = CollectionAdminRequest.deleteCollection(collection)
-        .processAndWait(client, MAX_TIMEOUT_SECONDS);
-    assertSame("DeleteCollection did not complete", RequestStatusState.COMPLETED, state);
+    if (!legacy) {
+      state = CollectionAdminRequest.deleteCollection(collection)
+          .processAndWait(client, MAX_TIMEOUT_SECONDS);
+      assertSame("DeleteCollection did not complete", RequestStatusState.COMPLETED, state);
+    }
   }
 
   public void testAsyncIdRaceCondition() throws Exception {
@@ -252,7 +262,7 @@ public class CollectionsAPIAsyncDistributedZkTest extends SolrCloudTestCase {
               numSuccess.incrementAndGet();
             } catch (SolrServerException e) {
               if (log.isInfoEnabled()) {
-                log.info("Exception during collection reloading, we were waiting for one: ", e);
+                log.info(e.getMessage());
               }
               assertEquals("Task with the same requestid already exists.", e.getMessage());
               numFailure.incrementAndGet();
@@ -272,4 +282,17 @@ public class CollectionsAPIAsyncDistributedZkTest extends SolrCloudTestCase {
       }
     }
   }
+  
+  public void testAsyncIdBackCompat() throws Exception {
+    //remove with Solr 9
+    cluster.getZkClient().makePath("/overseer/collection-map-completed/mn-testAsyncIdBackCompat", true, true);
+    try {
+      CollectionAdminRequest.createCollection("testAsyncIdBackCompat","conf1",1,1)
+      .processAsync("testAsyncIdBackCompat", cluster.getSolrClient());
+      fail("Expecting exception");
+    } catch (SolrServerException e) {
+      assertTrue(e.getMessage().contains("Task with the same requestid already exists"));
+    }
+  }
+
 }
